requirements.txt
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
google-api-core==2.12.0
google-auth==2.23.0
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
idna==3.4
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
shellingham==1.5.3
typer==0.9.0
typing_extensions==4.8.0
urllib3==1.26.16
